Moved ae2 integration files into their own respective folders
