+++ dependencies.gradle
@@ -95,6 +95,4 @@ dependencies {
     compileOnly rfg.deobf('curse.maven:better-storage-232919:2248061')
     compileOnly rfg.deobf('curse.maven:compact-storage-223703:2276736')
     compileOnly rfg.deobf('curse.maven:actually-additions-228404:2274681')
-
-    testImplementation("junit:junit:4.13.2")
 }
