+++ src/main/java/com/cleanroommc/bogosorter/BogoSorter.java
@@ -17,13 +17,13 @@ import com.cleanroommc.bogosorter.common.SortConfigChangeEvent;
 import com.cleanroommc.bogosorter.common.XSTR;
 import com.cleanroommc.bogosorter.common.config.BogoSortCommandTree;
 import com.cleanroommc.bogosorter.common.config.Serializer;
-import com.cleanroommc.bogosorter.common.config.TooltipFeatureConfig;
+import com.cleanroommc.bogosorter.common.config.ae2.TooltipFeatureConfig;
 import com.cleanroommc.bogosorter.common.dropoff.DropOffButtonHandler;
 import com.cleanroommc.bogosorter.common.dropoff.DropOffScheduler;
-import com.cleanroommc.bogosorter.common.network.Ae2AmountService;
+import com.cleanroommc.bogosorter.common.network.ae2.Ae2AmountService;
 import com.cleanroommc.bogosorter.common.network.NetworkHandler;
 import com.cleanroommc.bogosorter.common.network.NetworkUtils;
-import com.cleanroommc.bogosorter.common.network.STooltipFeatureState;
+import com.cleanroommc.bogosorter.common.network.ae2.STooltipFeatureState;
 import com.cleanroommc.bogosorter.common.refill.RefillHandler;
 import com.cleanroommc.bogosorter.common.sort.ButtonHandler;
 import com.cleanroommc.bogosorter.common.sort.DefaultRules;
