+++ src/main/java/com/cleanroommc/bogosorter/common/network/ae2/Ae2Status.java
@@ -1,4 +1,4 @@
-package com.cleanroommc.bogosorter.common.network;
+package com.cleanroommc.bogosorter.common.network.ae2;
 
 public final class Ae2Status {
