+++ src/main/java/com/cleanroommc/bogosorter/common/network/ae2/CAe2AmountBatchRequest.java
@@ -1,16 +1,18 @@
-package com.cleanroommc.bogosorter.common.network;
+package com.cleanroommc.bogosorter.common.network.ae2;
 
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.cleanroommc.bogosorter.common.network.IPacket;
+import com.cleanroommc.bogosorter.common.network.NetworkUtils;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.NetHandlerPlayServer;
 import net.minecraft.network.PacketBuffer;
 import net.minecraftforge.fluids.FluidStack;
 
-import com.cleanroommc.bogosorter.common.config.TooltipFeatureConfig;
+import com.cleanroommc.bogosorter.common.config.ae2.TooltipFeatureConfig;
 
 public class CAe2AmountBatchRequest implements IPacket {
