+++ src/main/java/com/cleanroommc/bogosorter/common/network/ae2/STooltipFeatureState.java
@@ -1,7 +1,8 @@
-package com.cleanroommc.bogosorter.common.network;
+package com.cleanroommc.bogosorter.common.network.ae2;
 
 import java.io.IOException;
 
+import com.cleanroommc.bogosorter.common.network.IPacket;
 import net.minecraft.client.network.NetHandlerPlayClient;
 import net.minecraft.network.PacketBuffer;
